// web/src/lib/api/generated/models/internal-customers-customer-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface InternalCustomersCustomerResponse {
    'address'?: string;
    'created_at'?: string;
    'email'?: string;
    'id'?: string;
    'name'?: string;
    'phone'?: string;
    'updated_at'?: string;
}